scratch/requirements.txt
numpy
tqdm~=4.64.1
transformers~=4.22.2
jax~=0.3.21
torchvision~=0.13.0
pillow~=9.2.0
pyyaml~=6.0